refactor: move useAnimationTransition to src/hooks
