+++ src/components/Modal/ReanimatedModal/Backdrop/index.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import Animated, {Keyframe, ReduceMotion} from 'react-native-reanimated';
 import type {BackdropProps} from '@components/Modal/ReanimatedModal/types';
-import useAnimationTransition from '@components/Modal/ReanimatedModal/useAnimationTransition';
+import useAnimationTransition from '@hooks/useAnimationTransition';
 import {getModalInAnimation, getModalOutAnimation} from '@components/Modal/ReanimatedModal/utils';
 import {PressableWithoutFeedback} from '@components/Pressable';
 import useLocalize from '@hooks/useLocalize';
